use an abstract class
